refactor(imperva): extract captcha module + missing test coverage
bypass.rs was carrying the CAPTCHA challenge/solution types plus
both CDP helpers (extract_captcha_site_key, inject_captcha_solution).
That's ~140 lines of CAPTCHA-specific surface in a file whose job
is the poll loop. Split into a sibling captcha module:

- CaptchaChallenge / CaptchaSolution / CaptchaSolver type alias move
  to captcha.rs (re-exported from lib.rs unchanged for users).
- The two CDP helpers move with them.
- New arc_solver helper isolates the Arc<dyn Fn -> BoxFuture> wrap
  so bypass.rs's on_captcha just calls arc_solver(f).
- Local BoxFuture<'a, T> type alias dropped in favor of
  futures::future::BoxFuture (already a workspace dep).

Adds five tests for the captcha module that were missing from the
initial implementation:

- extract returns (None, url) when neither sitekey attr is present
- extract uses rcap for Recaptcha kind (and ignores hcap)
- extract returns None for ImpervaNative kind
- inject escapes both \ and " in form_field (regression test for
  the JS-escaping fix in the previous commit)
- inject returns JsError when the evaluator raises
